up1029

+++ src/main/java/org/example/Main.java
@@ -6,8 +6,8 @@ public class Main {
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long data = sc.nextLong();
+        double data = sc.nextDouble();
 
-        System.out.printf("%d", data);
+        System.out.printf("%.11f",data);
     }
 }
